Implement wrap_contents.sh script
This commit implements the wrap_contents.sh script. The script succesfully wraps the contents of input file between the header and footer specified in the second argument and puts the result in a file with the name specified by the third   argument. This commit also contains the output generated when you run the script as such: ./wrap_contents.sh meats.txt bread my_chart.html