Fix regex.sh script to work with r0
This commit fixes the regex.sh script so that it works with r0. The problem was that I assumed that the script would take in the desired r*_input.txt file and then output the matching r*_output.txt file. However, this assumption was incorrect. We are supposed to write the script such that we just call it once and then get all 3 r*_output.txt files. Since thetests were calling the script without any arguments, the code in the if statement never ran and thus the r0_output.txt file was never created.